+++ .env.example
@@ -8,3 +8,5 @@ UPSTASH_TOKEN = # required if APL is upstash
 ENCRYPT_KEY=my_key
 HYPERSWITCH_PROD_BASE_URL = # hyperswitch prod server url
 HYPERSWITCH_SANDBOX_BASE_URL =  # hyperswitch testing server url
+JUSPAY_SANDBOX_BASE_URL= # juspay testing server url
+JUSPAY_PROD_BASE_URL= # juspay prod server url
