+++ src/lib/env.mjs
@@ -29,7 +29,8 @@ export const env = createEnv({
     APP_IFRAME_BASE_URL: z.string().optional(),
     HYPERSWITCH_SANDBOX_BASE_URL: z.string(),
     HYPERSWITCH_PROD_BASE_URL: z.string(),
-    JUSPAY_BASEURL: z.string(),
+    JUSPAY_SANDBOX_BASE_URL: z.string(),
+    JUSPAY_PROD_BASE_URL: z.string()
   },
 
   /*
@@ -54,7 +55,6 @@ export const env = createEnv({
     ENV: process.env.ENV,
     NEXT_PUBLIC_SENTRY_DSN: process.env.NEXT_PUBLIC_SENTRY_DSN,
     NEXT_PUBLIC_ENV: process.env.ENV,
-    JUSPAY_BASEURL: process.env.JUSPAY_BASEURL,
     SECRET_KEY: process.env.SECRET_KEY,
     ENCRYPT_KEY: process.env.ENCRYPT_KEY,
     SENTRY_DSN: process.env.SENTRY_DSN,
@@ -71,5 +71,7 @@ export const env = createEnv({
     APP_IFRAME_BASE_URL: process.env.APP_IFRAME_BASE_URL,
     HYPERSWITCH_SANDBOX_BASE_URL: process.env.HYPERSWITCH_SANDBOX_BASE_URL,
     HYPERSWITCH_PROD_BASE_URL: process.env.HYPERSWITCH_PROD_BASE_URL,
+    JUSPAY_SANDBOX_BASE_URL: process.env.JUSPAY_SANDBOX_BASE_URL,
+    JUSPAY_PROD_BASE_URL: process.env.JUSPAY_PROD_BASE_URL
   },
 });
